Adjust Windows Installation Instructions to account for Nuclio issue#1821 (#5558)
In my understanding of https://github.com/nuclio/nuclio/issues/1821, the
Nuctl (1.8.14) CLI is looking for a path that is only valid on a Linux
environment, which it does not find when running via Git Bash (even when
using the Windows version of Nuctl). However, installing CVAT onto a
Linux VM allows Nuctl to locate this path and operate normally. I
initially found this when setting up CVAT myself on Git Bash as per the
given instructions for Windows 10.

(I am still learning how to use GitHub as far as pull requests / forks /
etc work, sorry if this is not the right way to approach this change.
Please let me know if I've missed something important.)

### How has this been tested?
This is only a change to instructions, but I did test this on multiple
machines . As long as the machine is capable of running a Linux kernel
it shouldn't run into any issues.