Fix typing of credentials parameter of make_client in the Python SDK (#5609)
### Motivation and context
This is a very simple pull request. The type of the credentials
parameter of `make_client` is currently `Optional[Tuple[int, int]]`, but
it should be `Optional[Tuple[str, str]]` as used by `Client#login`. This
PR makes that change.

### How has this been tested?
The typing does not affect the functionality of the code (just the
warnings I get in an IDE).