Fix skeleton tracks (#6075)
Currently, we don't have validation of incoming annotations, as a
result, there is exist some cases when annotations successfully saved in
DB, but it's impossible to export them. In order to successfully export
a dataset with a skeleton track it's required that each track satisfy
the following condition:
` {frame number of track} == {frame number of parent track} == {frame
number of the first shape of the track}`

This PR adds an additional step during saving annotation in DB. This
additional step check that all these there "frame numbers" are equal and
try to automatically fix it's not true.